添加本地构建

+++ build.ps1
@@ -1,4 +1,6 @@
-Remove-Item .\_site -Recurse -Force
+if (test-path .\_site) {
+  Remove-Item .\_site -Recurse -Force
+}
 # generate TOC
 DocFxTocGenerator -d ./zh/ater.web  -r -g -s
 DocFxTocGenerator -d ./zh/droplet  -r -g -s
